refactor some helpers out of views.py
